int64 indexing: fix loop counters and size variables
- Default.MatMul.cs: for (int i -> for (long i when iterating over len
- Default.Round.cs: for (int i -> for (long i for element loops
- np.repeat.cs: int outIdx/i -> long for srcSize iteration
- NDArray.negative.cs: for (int i -> for (long i for element loops
- np.random.gamma.cs: for (int i -> for (long i in Marsaglia method
- NDArray.Indexing.Selection.Getter.cs: int i -> long i for dst.size
- NDArray.Indexing.Selection.Setter.cs: int i -> long i for len loop
- np.nanmean/std/var.cs: comprehensive migration
  - List<int> -> List<long> for output shape
  - var outputSize = 1 -> long outputSize = 1
  - var axisLen -> long axisLen
  - for (int outIdx -> for (long outIdx
  - for (int k -> for (long k for axis iteration
  - int temp -> long temp for coordinate conversion
  - Added overflow check for .NET array indexing
- Arrays.cs: for (int i -> for (long i in long[] Slice method

+++ src/NumSharp.Core/Backends/Default/Math/Default.Round.cs
@@ -37,19 +37,19 @@ namespace NumSharp.Backends
                     case NPTypeCode.Double:
                     {
                         var out_addr = (double*)@out.Address;
-                        for (int i = 0; i < len; i++) out_addr[i] = Math.Round(out_addr[i], decimals);
+                        for (long i = 0; i < len; i++) out_addr[i] = Math.Round(out_addr[i], decimals);
                         return @out;
                     }
                     case NPTypeCode.Single:
                     {
                         var out_addr = (float*)@out.Address;
-                        for (int i = 0; i < len; i++) out_addr[i] = (float)Math.Round(out_addr[i], decimals);
+                        for (long i = 0; i < len; i++) out_addr[i] = (float)Math.Round(out_addr[i], decimals);
                         return @out;
                     }
                     case NPTypeCode.Decimal:
                     {
                         var out_addr = (decimal*)@out.Address;
-                        for (int i = 0; i < len; i++) out_addr[i] = decimal.Round(out_addr[i], decimals);
+                        for (long i = 0; i < len; i++) out_addr[i] = decimal.Round(out_addr[i], decimals);
                         return @out;
                     }
                     default:
